+++ src/simulationBox/celllist.cpp
@@ -255,13 +255,7 @@ void CellList::addMoleculesToCells(SimulationBox &simulationBox)
             const auto atomCellIndices = getCellIndexOfAtom(box, position);
             const auto cellIndexScalar = getCellIndex(atomCellIndices);
 
-            const auto &[_, successful] = mapCellIndexToAtomIndex.try_emplace(
-                cellIndexScalar,
-                std::vector<size_t>({j})
-            );
-
-            if (!successful)
-                mapCellIndexToAtomIndex[cellIndexScalar].push_back(j);
+            mapCellIndexToAtomIndex[cellIndexScalar].push_back(j);
         }
 
         auto addMoleculeAndAtomIndicesToCell = [this, molecule](auto &pair)